+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @elf-framework/design-system
 
+## 0.0.134
+
+### Patch Changes
+
+- add compact property in CardBody component
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.134
+
+## 0.0.133
+
+### Patch Changes
+
+- update panel style
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.133
+
+## 0.0.132
+
+### Patch Changes
+
+- update panel style
+
+## 0.0.131
+
+### Patch Changes
+
+- update panel style
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.131
+
 ## 0.0.128
 
 ### Patch Changes